Create a globally-shared TextualHelper module
.. and move our new generic helpers into it.
